[rt.cpan.org #75474] Problems with UTF8 in folder- and file-names
The documentation mentions UTF-8 in *data*, but not in file or folder
names, and the script adds files as-is.

The problem as described was deduced from a bigger project where the
file names came from a utf8 encode database. further experiments lead
to a solution

 $tar->add_files (encode ("utf-8", $f));

fixed all. This might be explained in the documentation (under a
specific header where file- and foldernames are subject) or might even
be done automatically in the add_files () method when the target file
is marked utf-8

Patch encodes filenames automatically if marked utf-8

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>