BUG in Archive-Tar: chown won't work on symlinks; patch included
I had trouble using Archive::Tar (version 1.36) on Linux (Ubuntu 8.04);
however, the issue is not related to this specific version of
Archive::Tar nor the Linux distribution.

If I untar a directory (created by Archive::Tar) that has a symlink
pointing to a file within the same directory, then sometime the links
seems to be extracted *before* the file it points to. Therefore chown()
fails with:

Could not set uid/gid on '/usr/home/gateway/CURRENT'

Looking at the code, neither utime() nor chmod() are called for symlink,
but chown() is called also for links even though it should be handled
similar to chmod().
Or is there any reason to react differently for chmod() and chown()?

A patch for 1.86 would be:

--- Archive-Tar-1.86/lib/Archive/Tar.pm 2012-05-24 13:34:49.000000000 +0200
+++ Archive-Tar-1.86/lib/Archive/Tar_new.pm     2012-06-01
11:43:22.348225000 +0200
@@ -874,7 +874,7 @@
              $self->_error( qq[Could not update timestamp] );
      }

-    if( $CHOWN && CAN_CHOWN->() ) {
+    if( $CHOWN && CAN_CHOWN->() and not -l $full ) {
          chown $entry->uid, $entry->gid, $full or
              $self->_error( qq[Could not set uid/gid on '$full'] );
      }

(It would of course be possible to build a big "if (not -l $full) {"
around utime(), chown() and chmod())

Besides this issue, Archive::Tar works great! Thanks for your work!

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>