Fix pod in bin/ptar (#12)
Several of the pod paragraphs in this file are indented, making them
verbatim instead of the intended flowing text.

This would cause, for example the L<> link to be displayed not as a link
but as the characters "L" "<" ... ">".  And is is not proper to have a
verbatim paragraph in a NAME section.

This commit merely outdents things so they are no longer considered
verbatim, and creates a link to the tar man page that is referenced.