Fix absolute path handling on VMS.
a00e0305c16256c6f broke absolute path handling on filesystems that
have the concept of a volume because it discarded the volume
returned from splitpath.  This restores that on VMS only.

Also, to get the absolute path test added in 53dac726398c2c08
working on VMS, it's necessary to remove the root directory name
(000000) and ensure all path components are in Unix format.