Archive::Tar: Generalize for EBCDIC
The only real change to get Archive::Tar to work on EBCDIC is that the
bzip magic number is based on ASCII characters, and Archive::Tar was
using native values (hence EBCDIC on EBCDIC platforms).  Simply change
this to expect the ASCII values, and everything works.

However, the test files were also changed to skip tests concerning files
shipped with the distribution that were encoded in ASCII.  These would
need, once unarchived, to be translated to EBCDIC for proper comparison.
It's simpler to just skip them.

Files that are archived and unarchived on platforms with the same
character set, or files that contain only binary data work fine.