Update links under SEE ALSO
I was reading the perldoc via perldoc.perl.org and noticed that the
links in SEE ALSO were not clickable; so I fixed that.

Then I noticed that some of the links in SEE ALSO were very non-existing
anymore, and I guess at this point the difference between GNU tar and
POSIX is not super interesting anymore because GNU tar has been POSIX
compliant for 10+ years at this point? So I removed those links.

But if you want they could probably also be replaced with pointers to
archive.org or such.